add soup_utils.py
